requirements.txt
SQLAlchemy==2.0.35
aiogram>=3.4.1
alembic==1.13.2
email-validator==2.2.0
fastapi==0.115.0
httpx==0.27.0
httpx>=0.27.0
jinja2==3.1.4
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.9
pydantic==2.9.2
pyjwt==2.9.0
python-multipart==0.0.9
redis==5.0.8
redis>=5.0.1
uvicorn[standard]==0.30.6
openpyxl>=3.1.2


